pyproject.toml
[tool.poetry]
name = "nautobot-ssot-vsphere"
version = "0.1.0"
description = "Nautobot SSoT vSphere"
authors = ["h4ndzdatm0ld <hugotinoco@icloud.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/h4ndzdatm0ld/nautobot-ssot-vsphere"
repository = "https://github.com/h4ndzdatm0ld/nautobot-ssot-vsphere"
keywords = ["nautobot", "nautobot-plugin", "ssot", "diffsync", "vsphere", "vmware", "virtualization"]
include = [
    "LICENSE",
    "README.md",
]
packages = [
    { include = "nautobot_ssot_vsphere" },
]

[tool.poetry.dependencies]
# Used for local development
nautobot = { version = "^1.2.0", optional = true }
python = "^3.7"
diffsync = "^1.4.3"
nautobot-ssot = "^1.1.0"

[tool.poetry.dev-dependencies]
bandit = "*"
black = "*"
coverage = "*"
django-debug-toolbar = "*"
flake8 = "*"
invoke = "*"
ipython = "*"
mkdocs = "*"
pydocstyle = "*"
pylint = "*"
pylint-django = "*"
pytest = "*"
yamllint = "*"
Markdown = "*"
Sphinx = "^4.5.0"
myst-parser = "^0.17.0"
sphinx-rtd-theme = "^1.0.0"
sphinx-autoapi = "^1.8.4"
jedi = "^0.18.1"
toml = "^0.10.2"

[tool.poetry.extras]
nautobot = ["nautobot"]

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''

[tool.pylint.master]
# Include the pylint_django plugin to avoid spurious warnings about Django patterns
load-plugins="pylint_django"
ignore=".venv"

[tool.pylint.basic]
# No docstrings required for private methods (Pylint default), or for test_ functions, or for inner Meta classes.
no-docstring-rgx="^(_|test_|Meta$)"

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
disable = """,
    line-too-long,
    bad-continuation,
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """

[tool.pydocstyle]
convention = "google"
inherit = false
match = "(?!__init__).*\\.py"
match-dir = "(?!tests|migrations|development)[^\\.].*"
# D212 is enabled by default in google convention, and complains if we have a docstring like:
# """
# My docstring is on the line after the opening quotes instead of on the same line as them.
# """
# We've discussed and concluded that we consider this to be a valid style choice.
add_ignore = "D212"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
addopts = "-vv --doctest-modules"

[tool.isort]
profile = "black"
